Move CapacityResource to resource package
Accidentally ended up in the wrong package in #211. This caused it to be
omitted from the `api` profile.

Testing
-------
Run using `SPRING_PROFILES_ACTIVE=api RBAC_USE_STUB=true ./gradlew
bootRun`.

Try:

```
curl -v "http://localhost:8080/api/rhsm-subscriptions/v1/capacity/products/RHEL?granularity=daily&beginning=2017-07-21T17%3A32%3A28Z&ending=2017-07-21T17%3A32%3A28Z" -H  "accept: application/vnd.api+json" -H  "x-rh-identity: eyJpZGVudGl0eSI6eyJhY2NvdW50X251bWJlciI6ImFjY291bnQxMjMiLCJ0eXBlIjoiVXNlciIsInVzZXIiOnsiaXNfb3JnX2FkbWluIjp0cnVlfSwiaW50ZXJuYWwiOnsib3JnX2lkIjoib3JnMTIzIn19fQo"
```

Without the change you get a 404. With you get a 200.
